+++ tests/pytest_tests/unit_tests/test_launch/test_controller/test_base.py
@@ -1,9 +1,8 @@
 from unittest.mock import AsyncMock, MagicMock
 
 import pytest
-from wandb.sdk.launch.agent2.controller import LaunchControllerConfig
 from wandb.sdk.launch.agent2.controllers.base import BaseManager
-from wandb.sdk.launch.agent2.jobset import Job, JobSetSpec, create_jobset
+from wandb.sdk.launch.agent2.jobset import Job
 from wandb.sdk.launch.queue_driver.abstract import AbstractQueueDriver
 
 
@@ -19,62 +18,6 @@ class TestBaseManager(BaseManager):
 
     async def find_orphaned_jobs(self):
         pass
-
-
-@pytest.fixture
-def jobset_spec():
-    return JobSetSpec(name="test", entity_name="test-entity", project_name="test-proj")
-
-
-@pytest.fixture
-def jobset(jobset_spec):
-    entity = "test-entity"
-    mock_api = AsyncMock()
-    mock_api.get_jobset_by_spec.return_value = {
-        "jobSetDiff": {
-            "jobs": [
-                {
-                    "createdAt": "2024-04-09T16:33:45",
-                    "updatedAt": None,
-                    "id": "test-id",
-                    "runSpec": {
-                        "job": f"{entity}/test-proj/job-ktest12abc:latest",
-                        "author": f"{entity}",
-                        "entity": f"{entity}",
-                        "project": "test-proj",
-                        "resource": "local-container",
-                        "overrides": {"args": [], "run_config": {}, "entry_point": []},
-                        "resource_args": {"local-container": {}},
-                        "_wandb_job_collection_id": "arti-id",
-                    },
-                    "priority": 2,
-                    "state": "PENDING",
-                    "associatedRunId": None,
-                    "launchAgentId": None,
-                }
-            ],
-            "metadata": {
-                "@target_resource": "local-container",
-                "@id": "jobset-id",
-                "@name": "docker-queue-kyle",
-                "@capacity": 500,
-                "@max_concurrency": "auto",
-                "@prioritization_mode": "V0",
-            },
-        }
-    }
-    return create_jobset(jobset_spec, MagicMock(), "test-agent", MagicMock())
-
-
-@pytest.fixture
-def controller_config(jobset, jobset_spec):
-    return LaunchControllerConfig(
-        {
-            "agent_id": jobset.agent_id,
-            "jobset_spec": jobset_spec,
-            "jobset_metadata": jobset._metadata,
-        }
-    )
 
 
 @pytest.fixture
